Enhancements of the HTTP server.
Added HTTP request line and header reading.
More data is now read directly from the reader functions so
avoided read redundancy.
Server returns now bad request if packet is incomplete and no more
data is available to read.
The backlog value of the listen socket is changed back to the default
one of 5, because there is no obvious profit of that.