Refactoring and some modification.
Environment is now read from the http_server:start_link function.
Added setter for the log handler to the http_server (http_handler is
updated accordignly).
Fixed crash on stopping not listening server.
Config is now part of state, so could be accessible when needed.
Separate configuration file is now not supported temporarily.
Refactoring is provided.